microservices/imagizer/requirements.txt
fastapi == 0.85.0
pillow == 9.2.0
uvicorn == 0.18.3
requests == 2.28.1
starlette-exporter == 0.14.0
boto3 == 1.24.89